feat(app): add imprint and app information
Add an app information area to the app which will provide the user with
all necessary information. At the moment the imprint is in place as well
as a link to the apps GitHub repository and the current version of the
app is shown.
Additionally the about section on the start page explains the goal of
the app and links to the app information area.